requirements.txt
ajax==3.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
async-timeout==4.0.2
captcha==0.4
decorator==5.1.1
Django==4.2.2
django-appconf==1.0.5
django-ranged-response==0.2.0
djangorestframework==3.14.0
Pillow==9.5.0
python-bidi==0.4.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
